backend/requirements.txt
# ── Django Core ───────────────────────────────────────────────────────────
Django>=5.0,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.3,<5.0
django-environ>=0.11,<1.0
django-filter>=24.0,<25.0

# ── Auth ──────────────────────────────────────────────────────────────────
djangorestframework-simplejwt>=5.3,<6.0
django-allauth[socialaccount]>=0.61,<1.0

# ── API Docs ──────────────────────────────────────────────────────────────
drf-spectacular>=0.27,<1.0

# ── Database ──────────────────────────────────────────────────────────────
psycopg2-binary>=2.9,<3.0

# ── Cache & Queue ─────────────────────────────────────────────────────────
redis>=5.0,<6.0
celery>=5.3,<6.0
django-celery-beat>=2.6,<3.0
django-celery-results>=2.5,<3.0

# ── Stripe — dj-stripe 2.9.x requires stripe>=8.0,<12.0 ──────────────────
dj-stripe>=2.9,<2.10
stripe>=8.0,<12.0

# ── File Storage ─────────────────────────────────────────────────────────
django-storages>=1.14,<2.0
boto3>=1.34,<2.0

# ── PDF ───────────────────────────────────────────────────────────────────
weasyprint>=60.0,<63.0

# ── SMS ──────────────────────────────────────────────────────────────────
twilio>=8.0,<10.0

# ── Google Calendar ───────────────────────────────────────────────────────
google-auth>=2.0,<3.0
google-api-python-client>=2.0,<3.0

# ── Utilities ─────────────────────────────────────────────────────────────
Pillow>=10.0,<11.0
django-ratelimit>=4.0,<5.0

# ── Testing ──────────────────────────────────────────────────────────────
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
factory-boy>=3.3,<4.0
